# api/openapi.yaml
openapi: 3.1.0
info:
  title: Flex API
  version: 0.5.0
  description: API du serveur multimédia personnel Flex.
servers:
  - url: /api
security:
  - cookieAuth: []
paths:
...
  /users:
    get:
      operationId: listUsers
      summary: Liste les comptes Flex. Réservé aux administrateurs.
      responses:
        "200":
          description: Liste des utilisateurs.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AuthUser"
        "403":
          description: Droits administrateur requis.
    post:
      operationId: createUser
      summary: Crée un compte Flex. Réservé aux administrateurs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Utilisateur créé.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Nom d’utilisateur déjà utilisé.
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: updateUser
      summary: Modifie un compte Flex. Réservé aux administrateurs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Utilisateur modifié.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "409":
          description: Le dernier administrateur actif doit être conservé.
    delete:
      operationId: deleteUser
      summary: Supprime un compte et ses sessions. Réservé aux administrateurs.
      responses:
        "204":
          description: Utilisateur supprimé.
        "409":
          description: Le dernier administrateur actif doit être conservé.
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: flex_session
...
  schemas:
    AuthUser:
      type: object
      additionalProperties: false
      required: [id, username, role, active]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
...
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, password, role]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 12
          maxLength: 256
        role:
          type: string
          enum: [admin, user]
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required: [username, role, active]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
...